requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Transformer libraries
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Training and evaluation
wandb>=0.15.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Data handling
pandas>=2.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.23.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.4.0